Adding fixed include filters

+++ GPRO Net SDK/include/gpro-net/gpro-net-server/Physics_Update.h
@@ -5,7 +5,7 @@
 
 const float PHYSICS_GRAVITY = -9.81f;
 
-void updateVelocity(float vel[3], float accel[3], float dt);
-void updatePosition(float pos[3], float vel[3], float dt);
+void updateVelocity(float vel[3], float accel[3], double dt);
+void updatePosition(float pos[3], float vel[3], double dt);
 
 #endif // !PHYSICS_UPDATE_H
